setting up uploads

+++ src/modules/files/files.service.ts
@@ -1,11 +1,21 @@
-import { Injectable, Logger } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import {
+  Logger,
+  Injectable,
+  InternalServerErrorException,
+} from '@nestjs/common';
+
+import { v4 as uuid } from 'uuid';
 
 import { makeFileCacheKey } from './common/utils';
+
 import { GetFileDto } from './dtos/get-file.dto';
 import { UploadFileDto } from './dtos/upload-file.dto';
 import { UpdateFileDto } from './dtos/update-file.dto';
 
 import { MINUTES_10 } from '../../common/constants';
+import { S3Service } from '../../core/s3/s3.service';
+import { SqsService } from '../../core/sqs/sqs.service';
 import { RedisService } from '../../core/redis/redis.service';
 import { DatabaseService } from '../../core/database/database.service';
 
@@ -14,7 +24,10 @@ export class FilesService {
   logger = new Logger(FilesService.name);
 
   constructor(
+    private readonly s3Service: S3Service,
+    private readonly sqsService: SqsService,
     private readonly redisService: RedisService,
+    private readonly configService: ConfigService,
     private readonly databaseService: DatabaseService,
   ) {}
 
@@ -23,15 +36,29 @@ export class FilesService {
     dto: UploadFileDto,
     userId: string,
   ) {
-    console.log('the data', dto);
+    const key = uuid();
 
-    //FIXME: UPLOAD THE FILE TO S3
+    const { success, error } = await this.s3Service.uploadToS3({
+      Key: key,
+      Body: file.buffer,
+      Tagging: `lifetime=${dto.lifetime}`,
+      Bucket: this.configService.getOrThrow('S3_BUCKET_NAME'),
+    });
+
+    if (!success) {
+      this.logger.error({
+        error,
+        message: 'Failed to upload file to s3',
+      });
+
+      throw new InternalServerErrorException();
+    }
 
     await this.databaseService.files.create({
       data: {
-        description: dto.description,
-        s3_key: '', //FIXME:
+        s3_key: key,
         user_id: userId,
+        description: dto.description,
       },
     });
   }
@@ -89,7 +116,19 @@ export class FilesService {
   }
 
   async deleteSingleFile(userId: string, fileId: string) {
-    //FIXME: push the payload to sqs
+    const queued = await this.sqsService.pushMessage({
+      MessageBody: JSON.stringify({ userId, fileId }),
+      QueueUrl: this.configService.getOrThrow('SQS_QUEUE_URL'),
+    });
+
+    if (!queued.success) {
+      this.logger.error({
+        error: queued.error,
+        message: 'Failed to queue file for deletion',
+      });
+
+      throw new InternalServerErrorException();
+    }
 
     await this.databaseService.files.delete({
       where: {
@@ -112,7 +151,11 @@ export class FilesService {
     return { message: 'success' };
   }
 
-  async updateSingleFile(userId: string, fileId: string, dto: UpdateFileDto) {
+  async updateSingleFile(
+    userId: string,
+    fileId: string,
+    dto: UpdateFileDto,
+  ): Promise<GetFileDto> {
     const file = await this.databaseService.files.update({
       where: {
         id: fileId,
